+++ etc/upgrades/v26-gateway/stage-upgrade-info.yaml
