+++ zkstack_cli/crates/zkstack/Cargo.toml
@@ -65,3 +65,4 @@ xshell.workspace = true
 # Features that allows gateway-chain related actions.
 # These should be available for outside users until stabilized.
 gateway = []
+v27_evm_interpreter = ["gateway"]
