+++ zkstack_cli/crates/zkstack/src/commands/dev/mod.rs
@@ -1,10 +1,8 @@
 use clap::Subcommand;
 use commands::status::args::StatusArgs;
 use messages::MSG_STATUS_ABOUT;
-#[cfg(feature = "gateway")]
-use messages::{
-    MSG_GATEWAY_FINALIZE, MSG_GATEWAY_REGISTER_L2_TOKENS, MSG_GATEWAY_UPGRADE_CALLDATA,
-};
+#[cfg(feature = "v27_evm_interpreter")]
+use messages::MSG_V27_EVM_INTERPRETER_UPGRADE;
 use xshell::Shell;
 
 use self::commands::{
@@ -51,17 +49,9 @@ pub enum DevCommands {
     Status(StatusArgs),
     #[command(about = MSG_GENERATE_GENESIS_ABOUT, alias = "genesis")]
     GenerateGenesis,
-    #[cfg(feature = "gateway")]
-    #[command(about = MSG_GATEWAY_UPGRADE_CALLDATA)]
-    GatewayUpgradeCalldata(commands::gateway::GatewayUpgradeCalldataArgs),
-    #[cfg(feature = "gateway")]
-    #[command(about = MSG_GATEWAY_FINALIZE)]
-    GatewayUpgradeFinalization(
-        commands::gateway_finalize_preparation::GatewayFinalizePreparationArgs,
-    ),
-    #[cfg(feature = "gateway")]
-    #[command(about = MSG_GATEWAY_REGISTER_L2_TOKENS)]
-    GatewayL2TokenRegistration(commands::gateway_register_l2_tokens::GatewayRegisterL2TokensArgs),
+    #[cfg(feature = "v27_evm_interpreter")]
+    #[command(about = MSG_V27_EVM_INTERPRETER_UPGRADE)]
+    V27EvmInterpreterUpgradeCalldata(commands::v27_evm_eq::V27EvmInterpreterCalldataArgs),
 }
 
 pub async fn run(shell: &Shell, args: DevCommands) -> anyhow::Result<()> {
@@ -80,15 +70,9 @@ pub async fn run(shell: &Shell, args: DevCommands) -> anyhow::Result<()> {
         }
         DevCommands::Status(args) => commands::status::run(shell, args).await?,
         DevCommands::GenerateGenesis => commands::genesis::run(shell).await?,
-        #[cfg(feature = "gateway")]
-        DevCommands::GatewayUpgradeCalldata(args) => commands::gateway::run(shell, args).await?,
-        #[cfg(feature = "gateway")]
-        DevCommands::GatewayUpgradeFinalization(args) => {
-            commands::gateway_finalize_preparation::run(shell, args).await?
-        }
-        #[cfg(feature = "gateway")]
-        DevCommands::GatewayL2TokenRegistration(args) => {
-            commands::gateway_register_l2_tokens::run(args).await?
+        #[cfg(feature = "v27_evm_interpreter")]
+        DevCommands::V27EvmInterpreterUpgradeCalldata(args) => {
+            commands::v27_evm_eq::run(shell, args).await?
         }
     }
     Ok(())
